vim changes (w for up!, 2 for shrinkall), row spacing and attempt to not cause unlink bin error

+++ src/files-to-copy/.zshrc
@@ -62,7 +62,12 @@ plugins=(git)
 
 # export PATH=$HOME/bin:/usr/local/bin:$PATH
 # export MANPATH="/usr/local/man:$MANPATH"
-unlink ~/bin
+if [ -e ~/bin ]
+then
+else
+    echo 'bin should exist'
+    unlink ~/bin
+fi
 
 source $ZSH/oh-my-zsh.sh
